front/requirements.txt
Flask==3.0.3
grpcio==1.62.1